+++ controllers/saisieCom/ajoutIndController.js
@@ -1,0 +1,23 @@
+// commande_afficheNewInd, commande_creerNewInd
+const Individu = require('../../models/individu');
+
+const commande_afficheNewInd = (req, res) => {
+    res.render('./saisieCom/AjoutInd', { title: 'Commandes', style: "Commande" })
+}
+
+const commande_creerNewInd = (req, res) => {
+    const individu = new Individu(req.body);
+    individu.age = getAge(individu.dateNaissance)
+    individu.save()
+        .then((result) => {
+            res.redirect('/creerCom');
+        })
+        .catch((err) => {
+            console.log(err);
+        });
+}
+
+module.exports = {
+    commande_afficheNewInd,
+    commande_creerNewInd
+}
