+++ routes/adminRef/modifIndividuRoutes.js
@@ -1,45 +1,13 @@
 const express = require('express')
-const Individu = require('../../models/individu');
+const modifIndividuController = require('../../controllers/adminRef/modifIndividuController')
 
 const router = express.Router();
 
-// affiche liste de tous les individu de la base
-//ordonés avec celui ajouté le plus récemment en premier
-router.get('/', (req, res) => {
-    let searchOptions = {};
-    console.log(req.query);
-    if (req.query.nom != null && req.query.prenom != null && req.query.dateNaissance != null) {
-        if (req.query.dateNaissance != '') {
-            searchOptions.dateNaissance = req.query.dateNaissance;
-        }
-        searchOptions.nom = new RegExp(req.query.nom, 'i');
-        searchOptions.prenom = new RegExp(req.query.prenom, 'i');
-    }
-    console.log(searchOptions);
-    Individu.find(searchOptions).sort({ createdAt: -1 }).limit(10)
-        .then((result) => {
-            res.render('./adminRef/ModifIndividu', {
-                title: 'Administration du référentiel',
-                individus: result,
-                style: "Referentiel",
-                searchOptions: req.query
-            });
-        })
-        .catch((err) => {
-            console.log(err);
-        });
-});
+// affiche liste de tous les individu de la base, ordonnés avec celui ajouté le plus récemment en premier
+// Permet d'effectuer la recherche des individus en fonction du nom prénom date de naissance
+router.get('/', modifIndividuController.modifIndividu_recherche)
 
 // supprime l'individu sélectionné
-router.delete('/:id', (req, res) => {
-    const id = req.params.id;
-    Individu.findByIdAndDelete(id)
-        .then(result => {
-            res.json({ redirect: '/referentielModifIndividu' });
-        })
-        .catch((err) => {
-            console.log(err);
-        });
-});
+router.delete('/:id', modifIndividuController.modifIndividu_delete)
 
 module.exports = router;
