+++ routes/adminRef/referentielArticleRoutes.js
@@ -1,33 +1,13 @@
 const express = require('express')
-const Article = require('../../models/article');
+const referentielArticleController = require('../../controllers/adminRef/referentielArticleController')
 
 const router = express.Router();
 
 // affiche les informations d'un seul article sélectionné
 // dans la liste de recherche
-router.get('/:id', (req, res) => {
-    const id = req.params.id;
-    Article.findById(id)
-        .then(result => {
-            res.render('./adminRef/Article', { article: result, title: "Administration du référentiel", style: "Referentiel" });
-        })
-        .catch((err) => {
-            console.log(err);
-        });
-});
+router.get('/:id', referentielArticleController.article_getOne)
 
-router.put('/:id', async(req, res) => {
-    let article
-    try {
-        article = await Article.findById(req.params.id)
-        article.designation = req.body.designation
-        article.prix = req.body.prix
-        article.description = req.body.description
-        await article.save()
-        res.redirect('/referentielModifArticle')
-    } catch {
-        res.redirect('/referentiel')
-    }
-})
+// Enregistre les informations qui ont subi une modification
+router.put('/:id', referentielArticleController.article_modif)
 
 module.exports = router;
